refactor: reorganize file structure and rename files in frontend
