+++ frontend/src/App.tsx
@@ -1,54 +1,49 @@
-import {Layout, ConfigProvider, theme} from "antd";
 import {Route, Routes} from "react-router-dom";
-import ShiftGallery from "./antd/components/ShiftGallery.tsx";
-import useShiftState from "./antd/hooks/useShiftState.ts";
-import {useState} from "react";
+import {LoginForm} from "@/app/auth/LoginForm.tsx"
+import {ThemeProvider} from "./components/ThemeProvider.tsx";
+import {RegisterForm} from "@/app/auth/RegisterForm.tsx";
+import {ForgotPasswordForm} from "@/app/auth/ForgotPasswordForm.tsx";
+import ProtectedRoute from "@/app/auth/ProtectedRoute.tsx";
+import {AuthProvider} from "@/context/AuthProvider.tsx";
+import Dashboard from "@/app/dashboard/Dashboard.tsx";
 
 export default function App() {
-    const [isDarkMode, setIsDarkMode] = useState(true);
-    const toggleTheme = () => {
-        setIsDarkMode(prev => !prev);
-    };
-
-    const {
-        shiftList,
-        saveShift,
-        updateShift,
-        deleteShift
-    } = useShiftState();
-
-
     return (
+        <ThemeProvider defaultTheme="dark" storageKey="vite-ui-theme">
+            <AuthProvider>
+                <Routes>
+                    <Route path={"/login"} element={
+                        <div className="flex min-h-svh w-full items-center justify-center p-6 md:p-10">
+                            <div className="w-full max-w-sm">
+                                <LoginForm/>
+                            </div>
+                        </div>
+                    }/>
 
-        <ConfigProvider
-            theme={{
-                algorithm: isDarkMode ? theme.darkAlgorithm : theme.defaultAlgorithm,
-            }}>
-            <Layout style={{minHeight: "100vh"}}>
-                <Layout.Content
-                    style={{
-                        padding: "48px 24px",
-                        maxWidth: "1200px",
-                        width: "90%",
-                        margin: "0 auto",
-                        transition: "all 0.3s ease",
-                    }}
-                >
+                    <Route path={"/register"} element={
+                        <div className="flex min-h-svh w-full items-center justify-center p-6 md:p-10">
+                            <div className="w-full max-w-sm">
+                                <RegisterForm/>
+                            </div>
+                        </div>
+                    }/>
 
-                    <Routes>
-                        <Route path={"/"} element={
-                            <ShiftGallery
-                                shifts={shiftList}
-                                handleDelete={deleteShift}
-                                handleUpdate={updateShift}
-                                handleSave={saveShift}
-                                toggleTheme={toggleTheme}
-                                isDarkMode={isDarkMode}
-                            />
-                        }/>
-                    </Routes>
-                </Layout.Content>
-            </Layout>
-        </ConfigProvider>
-    );
+                    <Route path={"/recover"} element={
+                        <div className="flex min-h-svh w-full items-center justify-center p-6 md:p-10">
+                            <div className="w-full max-w-sm">
+                                <ForgotPasswordForm/>
+                            </div>
+                        </div>
+                    }/>
+
+                    <Route path="/home" element={
+                        <ProtectedRoute>
+                            <Dashboard/>
+                        </ProtectedRoute>
+                    }/>
+
+                </Routes>
+            </AuthProvider>
+        </ThemeProvider>
+    )
 }
